+++ legadilo/feeds/models/feed_article.py
@@ -6,6 +6,8 @@ from typing import TYPE_CHECKING
 
 from django.db import models
 
+from legadilo.utils.time_utils import utcnow
+
 if TYPE_CHECKING:
     from django_stubs_ext.db.models import TypedModelMeta
 else:
@@ -22,6 +24,9 @@ class FeedArticle(models.Model):
         "reading.Article", related_name="feed_articles", on_delete=models.PROTECT
     )
 
+    feed_article_id = models.TextField()
+    last_seen_at = models.DateTimeField(default=utcnow)
+
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
 
@@ -31,6 +36,12 @@ class FeedArticle(models.Model):
         constraints = [
             models.UniqueConstraint(
                 "feed", "article", name="%(app_label)s_%(class)s_article_linked_once_per_feed"
+            ),
+            # To speed up the search by feed_article_id
+            models.UniqueConstraint(
+                "feed",
+                "feed_article_id",
+                name="%(app_label)s_%(class)s_article_linked_once_per_feed_id",
             ),
         ]