+++ legadilo/feeds/models/feed_category.py
@@ -11,7 +11,7 @@ from django.db import models
 from django.utils.translation import gettext_lazy as _
 from slugify import slugify
 
-from legadilo.types import FormChoices
+from legadilo.utils.types import FormChoices
 
 if TYPE_CHECKING:
     from django_stubs_ext.db.models import TypedModelMeta