+++ legadilo/feeds/tests/test_models/test_feed.py
@@ -18,7 +18,7 @@ from legadilo.feeds.tests.factories import (
 )
 from legadilo.reading import constants as reading_constants
 from legadilo.reading.models import Article
-from legadilo.reading.tests.factories import ArticleFactory, TagFactory
+from legadilo.reading.tests.factories import ArticleDataFactory, ArticleFactory, TagFactory
 from legadilo.users.models import Notification
 from legadilo.users.tests.factories import UserFactory
 from legadilo.utils.testing import serialize_for_snapshot
@@ -329,7 +329,7 @@ class TestFeedManager:
             feed_url=ONE_ARTICLE_FEED_DATA.feed_url, user=user, disabled_at=utcnow()
         )
 
-        with django_assert_num_queries(16):
+        with django_assert_num_queries(18):
             feed, created = Feed.objects.create_from_metadata(
                 ONE_ARTICLE_FEED_DATA,
                 user,
@@ -345,7 +345,7 @@ class TestFeedManager:
         assert feed.feed_updates.count() == 1
 
     def test_create_from_feed_data(self, user, django_assert_num_queries):
-        with django_assert_num_queries(16):
+        with django_assert_num_queries(18):
             feed, created = Feed.objects.create_from_metadata(
                 FeedData(
                     feed_url="https://example.com/feeds/atom.xml",
@@ -407,7 +407,7 @@ class TestFeedManager:
     def test_create_from_metadata_with_tags(self, user, django_assert_num_queries):
         tag = TagFactory()
 
-        with django_assert_num_queries(19):
+        with django_assert_num_queries(21):
             feed, _ = Feed.objects.create_from_metadata(
                 ONE_ARTICLE_FEED_DATA,
                 user,
@@ -484,10 +484,12 @@ class TestFeedManager:
             main_source_type=reading_constants.ArticleSourceType.MANUAL,
             main_source_title="Not a feed",
             user=self.feed.user,
+            external_article_id="",
         )
-        FeedArticle.objects.create(feed=self.feed, article=existing_article)
+        with time_machine.travel("2025-01-01"):
+            FeedArticle.objects.create(feed=self.feed, article=existing_article)
 
-        with django_assert_num_queries(11):
+        with django_assert_num_queries(13), time_machine.travel("2025-07-01", tick=False):
             Feed.objects.update_feed(
                 self.feed,
                 FeedData(
@@ -543,15 +545,155 @@ class TestFeedManager:
         assert new_article.title == "Article 1"
         assert new_article.main_source_type == reading_constants.ArticleSourceType.FEED
         assert new_article.main_source_title == self.feed.title
+        assert new_article.external_article_id == "some-article-1"
+        new_article_feed_article = FeedArticle.objects.get(feed=self.feed, article=new_article)
+        assert new_article_feed_article.feed_article_id == "some-article-1"
+        assert new_article_feed_article.last_seen_at == utcdt(2025, 7, 1)
         existing_article.refresh_from_db()
         assert existing_article.main_source_type == reading_constants.ArticleSourceType.MANUAL
         assert existing_article.main_source_title != self.feed.title
+        assert not existing_article.external_article_id
+        existing_article_feed_article = FeedArticle.objects.get(
+            feed=self.feed, article=existing_article
+        )
+        assert existing_article_feed_article.feed_article_id == "some-article-existing"
+        assert existing_article_feed_article.last_seen_at == utcdt(2025, 7, 1)
+
+    def test_update_feed_some_articles_changed_url(self, django_assert_num_queries):
+        existing_article = ArticleFactory(
+            url="https://example.com/article/existing",
+            main_source_type=reading_constants.ArticleSourceType.FEED,
+            main_source_title="Some feed",
+            user=self.feed.user,
+        )
+        with time_machine.travel("2025-01-01"):
+            FeedArticle.objects.create(
+                feed=self.feed, article=existing_article, feed_article_id="some-article-existing"
+            )
+
+        with django_assert_num_queries(14), time_machine.travel("2025-07-01", tick=False):
+            Feed.objects.update_feed(
+                self.feed,
+                FeedData(
+                    feed_url="https://example.com/feeds/atom.xml",
+                    site_url="https://example.com",
+                    title="Awesome website",
+                    description="A description",
+                    feed_type=feeds_constants.SupportedFeedType.atom,
+                    etag="W/etag",
+                    last_modified=None,
+                    articles=[
+                        ArticleData(
+                            external_article_id="some-article-existing",
+                            title="Article 2",
+                            summary="Summary 2",
+                            content="Description existing updated",
+                            table_of_content=(),
+                            authors=("Author",),
+                            contributors=(),
+                            tags=(),
+                            url="http://example.com/new-url",
+                            preview_picture_url="",
+                            preview_picture_alt="",
+                            published_at=datetime.now(tz=UTC),
+                            updated_at=datetime.now(tz=UTC),
+                            source_title=self.feed.title,
+                            language="fr",
+                        ),
+                    ],
+                ),
+            )
+
+        assert self.feed.articles.count() == 1
+        assert self.feed.feed_updates.count() == 1
+        existing_article.refresh_from_db()
+        assert existing_article.url == "http://example.com/new-url"
+        existing_article_feed_article = FeedArticle.objects.get(
+            feed=self.feed, article=existing_article
+        )
+        assert existing_article_feed_article.feed_article_id == "some-article-existing"
+        assert existing_article_feed_article.last_seen_at == utcdt(2025, 7, 1)
+
+    def test_update_feed_with_republication(self, django_assert_num_queries):
+        existing_article_seen_recently = ArticleFactory(
+            url="https://example.com/article/seen-recently",
+            main_source_type=reading_constants.ArticleSourceType.FEED,
+            main_source_title="Some feed",
+            user=self.feed.user,
+            read_at=utcdt(2024, 6, 1),
+        )
+        FeedArticle.objects.create(
+            feed=self.feed,
+            article=existing_article_seen_recently,
+            feed_article_id="some-article-existing-seen-recently",
+            last_seen_at=utcdt(2025, 6, 1),
+        )
+        existing_article_read_recently = ArticleFactory(
+            url="https://example.com/article/read-recently",
+            main_source_type=reading_constants.ArticleSourceType.FEED,
+            main_source_title="Some feed",
+            user=self.feed.user,
+            read_at=utcdt(2025, 6, 1),
+        )
+        FeedArticle.objects.create(
+            feed=self.feed,
+            article=existing_article_read_recently,
+            feed_article_id="some-article-existing-read-recently",
+            last_seen_at=utcdt(2024, 6, 1),
+        )
+        existing_article_to_republish = ArticleFactory(
+            url="https://example.com/article/to-republish",
+            main_source_type=reading_constants.ArticleSourceType.FEED,
+            main_source_title="Some feed",
+            user=self.feed.user,
+            read_at=utcdt(2024, 6, 1),
+        )
+        FeedArticle.objects.create(
+            feed=self.feed,
+            article=existing_article_to_republish,
+            feed_article_id="some-article-existing-to-republish",
+            last_seen_at=utcdt(2024, 6, 1),
+        )
+
+        with django_assert_num_queries(14), time_machine.travel("2025-07-01", tick=False):
+            Feed.objects.update_feed(
+                self.feed,
+                FeedData(
+                    feed_url="https://example.com/feeds/atom.xml",
+                    site_url="https://example.com",
+                    title="Awesome website",
+                    description="A description",
+                    feed_type=feeds_constants.SupportedFeedType.atom,
+                    etag="W/etag",
+                    last_modified=None,
+                    articles=[
+                        ArticleDataFactory(
+                            external_article_id="some-article-existing-seen-recently"
+                        ),
+                        ArticleDataFactory(
+                            external_article_id="some-article-existing-read-recently"
+                        ),
+                        ArticleDataFactory(
+                            external_article_id="some-article-existing-to-republish"
+                        ),
+                    ],
+                ),
+            )
+
+        assert self.feed.articles.count() == 3
+        assert self.feed.feed_updates.count() == 1
+        existing_article_seen_recently.refresh_from_db()
+        assert existing_article_seen_recently.is_read
+        existing_article_read_recently.refresh_from_db()
+        assert existing_article_read_recently.is_read
+        existing_article_to_republish.refresh_from_db()
+        assert not existing_article_to_republish.is_read
 
     def test_update_feed_with_deleted_articles(self, django_assert_num_queries):
         deleted_url = "https://example.com/deleted/"
         FeedDeletedArticle.objects.create(article_url=deleted_url, feed=self.feed)
 
-        with django_assert_num_queries(10):
+        with django_assert_num_queries(12):
             Feed.objects.update_feed(
                 self.feed,
                 FeedData(