+++ legadilo/feeds/views/feeds_admin_view.py
@@ -14,8 +14,8 @@ from django.utils.translation import gettext_lazy as _
 from django.views.decorators.http import require_GET, require_http_methods
 
 from legadilo.feeds.models import Feed, FeedCategory, FeedTag
-from legadilo.types import FormChoices
 from legadilo.users.user_types import AuthenticatedHttpRequest
+from legadilo.utils.types import FormChoices
 
 from ...core.forms.fields import MultipleTagsField
 from ...reading.models import Tag