+++ legadilo/reading/views/tags_admin_views.py
@@ -18,9 +18,9 @@ from django.views.decorators.http import require_GET, require_http_methods
 from legadilo.core.forms.fields import MultipleTagsField, SlugifiableCharField
 from legadilo.reading.models import Tag
 from legadilo.reading.models.tag import SubTagMapping
-from legadilo.types import FormChoices
 from legadilo.users.models import User
 from legadilo.users.user_types import AuthenticatedHttpRequest
+from legadilo.utils.types import FormChoices
 
 
 @require_GET